pyproject.toml
[project]
name = "guestbook"
version = "0.1.0"
description = "Self-hosted RSVP website builder for home labs"
readme = "README.md"
authors = [
    { name = "Lawrence L. O'Boyle II", email = "vili.petal@protonmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "alembic>=1.13",
    "pydantic-settings>=2.0",
    "itsdangerous>=2.0",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "typer>=0.12",
    "qrcode[pil]>=7.4",
    "slowapi>=0.1",
    "email-validator>=2.0",
    "markdown>=3.10.2",
]

[project.optional-dependencies]
postgres = ["asyncpg>=0.29"]
dev = ["pytest>=8.0", "httpx>=0.27", "pytest-asyncio>=0.23"]

[project.scripts]
guestbook = "guestbook.cli:app"

[build-system]
requires = ["uv_build>=0.10.6,<0.11.0"]
build-backend = "uv_build"
